+++ CHANGELOG.md
@@ -18,12 +18,14 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * Added options to customize `pysat_ndtesting` instrument with sample rate,
     shift in time.
   * Added orbit number to `pysat_ndtesting`.
+  * Added the overwrite kwarg to `utils.registry.register_by_module`
 * Maintenance
   * Update link redirects in docs.
   * Improved Instrument ValueError messages.
   * Updated `Constellation.to_inst` method definition of coords, using dims
     to combine common dimensions instead.
   * Implement pyproject to manage metadata
+* Updated docstring references to `pysat.utils.files` in other modules.
   * Remove Sphinx cap
   * Add pandas cap
   * Remove deprecated `pysat_testing2d` instrument